+++ expression/builtin.go
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:generate sh -c "go run generator/*.go -outputdir ./"
+//go:generate go run generator/control_vec.go
 
 package expression
